Add files via upload
The temperature values are indipendent from the period of the day: may be any from 1500 to 9000. This behaviour let user change only the brightness instead of the colours of the monitors.
The -B option can be used with the curve from array_data: just use the option in the format 1.0:value:value.