Add files via upload
The temperature values are indipendent from the period of the day: may be any from 1500 to 9000. This behaviour let user change only the brightness instead of the colours of the monitors.
The -B option can be used with the curve from data_array: just use the option in the format 1.0:value:value.
The data_array file can be placed in HOME/.config/wlmonitorset directory or in this program folder.